refactor: Component name and API endpoint
