+++ src/views/ops/Template/components/SetVariableDialog.vue
